DateTimeTo24HOverStringConverterをIValueConverter実装に変更
XAMLからの使用は諦め, C#コードからの使用のみを想定することにした
